Add ability to disable opflex-agent features
The opflex-agent depends on features in the vSwitch from the
underlying distibution. Some distributions provide older versions
of the vSwitch, which lack support for features needed by the opflex-agent.
In order to ensure itneroperability with these distributions, a flag was
added to the opflex-agent to disable features that aren't supported by
the underlying distribution.

This adds the configuration of that list to the opflex-agent configuration.