Add OVS monitor
The docker container for the opflex-agent needs valid file handles/sockets
for OVS. If OVS, which is outside the container, gets restarted, the file
handles in the container are invalidated. This adds a monitor that checks
to see if the file handles exist, and if they don't, kills supervisord,
which will restart the container, and refresh the file handles.