Fix opflex-agent sequencing
The opflex-agent needs to wait until the OVS bridges that it manages
(br-int and br-fabric) exist before starting, or else it won't ever
know that the bridges exist, and therefore won't ever connect to them
and program flow-mods into their tables (it doesn't have the ability
to get async notifications from OVSDB). This patch adds configuration in
supervisord that causes it to wait until the bridges exist before
starting the opflex_agent.