Add opflex-agent restarts
Due to race conditions between the neutron-opflex-agent and the
opflex-agent restarts, we enxure they are sequenced by using the
restart feature in the opflex-agent. When a process in the neutron
opflex agent's container enters the RUNNING state, the opflex agent
is restarted after a 40 second delay, ensuring that there is no
conflict due to agent restarts.