Persist gen2 interface when droplog enabled
Rebooting the host results in the neutron-cleanup service getting
called, which removes interfaces from the ovs br-int bridge. The
gen2 interface used by droplog (and installed as part of the host-prep
steps in OSP) was getting deleted on reboots because of this.

This adds the creation of the gen2 interface to the opflex agent
container's supervisord configuration file, in order to ensure
that the gen2 interface gets recreated after reboots.